+++ NEWS.md
@@ -1,7 +1,9 @@
 # glebrt 0.4.0
+
 - Simplify `breaks_limits()` and allow more control over replacement of computed breaks
 
 # glebrt 0.3.0
+
 - Add `get_greys()` which returns up to five different shades of grey
 - Add functions for setting classes
   - `set_class()`: A pipeable version of `class(x) <-`